+++ iblqt/core.py
@@ -1,9 +1,12 @@
 """Non-GUI functionality, including event handling, data types, and data management."""
 
 import logging
+import sys
+import traceback
 import warnings
+from inspect import signature
 from pathlib import Path
-from typing import Any, cast
+from typing import Any, Callable, cast
 
 import numpy as np
 import numpy.typing as npt
@@ -16,6 +19,7 @@ from qtpy.QtCore import (
     QFileSystemWatcher,
     QModelIndex,
     QObject,
+    QRunnable,
     Qt,
     Signal,
     Slot,
@@ -767,3 +771,109 @@ class QAlyx(QObject):
         self._client.logout()
         self.statusChanged.emit(False)
         self.loggedOut.emit()
+
+
+class WorkerSignals(QObject):
+    """
+    Signals used by the :class:`Worker` class to communicate with the main thread.
+
+    Attributes
+    ----------
+    finished : Signal
+        Signal emitted when the worker has finished its task.
+
+    error : Signal(tuple)
+        Signal emitted when an error occurs. The signal carries a tuple with the exception type,
+        exception value, and the formatted traceback.
+
+    result : Signal(Any)
+        Signal emitted when the worker has successfully completed its task. The signal carries
+        the result of the task.
+
+    progress : Signal(int)
+        Signal emitted to report progress during the task. The signal carries an integer value.
+    """
+
+    finished = Signal()
+    error = Signal(tuple)
+    result = Signal(object)
+    progress = Signal(int)
+
+
+class Worker(QRunnable):
+    """
+    A generic worker class for executing functions concurrently in a separate thread.
+
+    This class is designed to run functions concurrently in a separate thread and emit signals
+    to communicate the results or errors back to the main thread.
+
+    Adapted from: https://www.pythonguis.com/tutorials/multithreading-pyqt-applications-qthreadpool/
+
+    Attributes
+    ----------
+    fn : Callable
+        The function to be executed concurrently.
+
+    args : tuple
+        Positional arguments for the function.
+
+    kwargs : dict
+        Keyword arguments for the function.
+
+    signals : WorkerSignals
+        An instance of WorkerSignals used to emit signals.
+
+    Methods
+    -------
+    run() -> None
+        The main entry point for running the worker. Executes the provided function and
+        emits signals accordingly.
+    """
+
+    def __init__(self, fn: Callable[..., Any], *args: Any, **kwargs: Any):
+        """
+        Initialize the Worker instance.
+
+        Parameters
+        ----------
+        fn : Callable
+            The function to be executed concurrently.
+
+        *args : tuple
+            Positional arguments for the function.
+
+        **kwargs : dict
+            Keyword arguments for the function.
+        """
+        super().__init__()
+        self.fn = fn
+        self.args = args
+        self.kwargs = kwargs
+        self.signals: WorkerSignals = WorkerSignals()
+        if 'progress_callback' in signature(fn).parameters:
+            self.kwargs['progress_callback'] = self.signals.progress
+
+    def run(self) -> None:
+        """
+        Execute the provided function and emit signals accordingly.
+
+        This method is the main entry point for running the worker. It executes the provided
+        function and emits signals to communicate the results or errors back to the main thread.
+
+        Returns
+        -------
+        None
+        """
+        try:
+            result = self.fn(*self.args, **self.kwargs)
+        except:  # noqa: E722
+            # Handle exceptions and emit error signal with exception details
+            traceback.print_exc()
+            exctype, value = sys.exc_info()[:2]
+            self.signals.error.emit((exctype, value, traceback.format_exc()))
+        else:
+            # Emit result signal with the result of the task
+            self.signals.result.emit(result)
+        finally:
+            # Emit the finished signal to indicate completion
+            self.signals.finished.emit()