+++ CHANGELOG.md
@@ -6,10 +6,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0]
+
+### Added
+- core.Worker: a generic worker class for executing functions concurrently in a 
+  separate thread
+
 ## [0.4.4] - 2025-03-26
 
 ### Fixed
-- compatibility issues with PySide in widgets.AlyxUserEdit and widgets.AlyxLoginWidget
+- compatibility issues with PySide in widgets.AlyxUserEdit and 
+  widgets.AlyxLoginWidget
 
 ### Changed
 - increased coverage of unit-tests 
@@ -107,6 +114,9 @@ _First release._
   providing color-mapped numerical data.
 - widgets.StatefulButton: A QPushButton that maintains an active/inactive state.
 
+[0.5.0]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.5.0
+[0.4.4]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.4.4
+[0.4.3]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.4.3
 [0.4.2]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.4.2
 [0.4.1]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.4.1
 [0.4.0]: https://github.com/int-brain-lab/iblqt/releases/tag/v0.4.0